Add files via upload
This Online food ordering system is a tiny project which sets up a food menu online and customers can easily place the order as per they like,with a food menu.The food ordering system like zomato Swiggy etc motivated me to develop the system like this. There are various Facilities provided so that the users of the system will get service effectively.